refactor: Move BlocklistService
BlocklistService is unrelated to (CoreMedia) contents. Thus, moving
outside of `content/` module and prefer flat module structure
instead of creating an additional layer.
