+++ packages/ckeditor5-coremedia-studio-integration/src/BlocklistServiceDescriptor.ts
